Consolidate transforms tests
